+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * The first formula in a workbook can now be a shared formula. [1223](https://github.com/JanMarvin/openxlsx2/pull/1223)
 * Avoid passing ASCII strings through `Rcpp::String()`. Previously all `cc` columns were passed through `Rcpp::String()` to avoid encoding issues on non unicode systems. [1224](https://github.com/JanMarvin/openxlsx2/pull/1224)
-
+* `fmt_txt()` is now indifferent about `color` and `colour`. [1225](https://github.com/JanMarvin/openxlsx2/pull/1225)
 
 ***************************************************************************